Resize everything
All the images were wayyyy to big before. This makes them all at most 300px, because that is the most they will be on the site. I think they still look good. Part of me wants to host the full resolution as well, and allow users to click and view that, but that can be for the future.